Modularize pool_controller pump logic and fix README docs
Split the monolithic __init__.py and pump_switch.{h,cpp} into
pump_anomaly.*, pump_current.*, pump_flow.*, and a schema/ package
(pump.py, heater.py, schedule.py, types.py) so each concern -
anomaly detection, current sensing, flow protection, and config
validation - lives in its own file.

Also corrects READMEs across all custom components where documented
options had drifted from the actual voluptuous schemas:
- pool_controller: removed the nonexistent enable_anomaly_detection
  option and documented the real anomaly_detection_switch,
  anomaly_status_binary_sensor, and anomaly_baseline_reset_button
  entities, plus the real on_anomaly trigger values.
- high_temp_water_heater: fixed a typo and documented the previously
  undocumented value ranges and cross-field validation rules.
- bed_sensor: documented default device_class/icon/accuracy_decimals
  values.
- treo_led_pool_light: documented that effects: is rejected.
- econet_zone_control: documented request_mod's range/"none" value,
  fan_modes minimum length, and the zone-lock/debounce behavior.
